Refactor: Organize webhooks and security decorators under app/bot
