+++ app/bot/decorators/__init__.py
@@ -1,0 +1,2 @@
+# app/bot/decorators/__init__.py
+# This file makes the 'decorators' directory under 'bot' a Python package.
