--- app/services/__init__.py
@@ -1,5 +1,0 @@
-"""Service layer for the application.
-
-This package contains modules that interact with external services
-and APIs, such as chat services (e.g., OpenAI, Azure OpenAI).
-"""
